feat: first whack at BLL for fetching Time Activities Details Report
* Also refactored (and sort of tested) `TimeActivityRepository.GetFiltered()` to use Dapper like the Customer and Employee filtered queries
* Done in service of #39

+++ src/Repositories/EmployeeRepository.cs
@@ -109,7 +109,7 @@ namespace DashAccountingSystemV2.Repositories
                  ON emp.""UserId"" = app_user.""Id""
    WHERE emp.""TenantId"" = @tenantId
      AND ( NOT @onlyActive OR entity.""Inactivated"" IS NULL ) -- TODO: Does emp.""ReleaseDate"" factor into this...?
-     AND ( @employeeNumbers::INTEGER[] IS NULL OR emp.""EmployeeNumber"" = ANY ( @employeeNumbers ) )
+     AND ( @employeeNumbers::BIGINT[] IS NULL OR emp.""EmployeeNumber"" = ANY ( @employeeNumbers ) )
 ORDER BY emp.""LastName""
         ,emp.""FirstName"";
 ";
